[Refactor]: command - entity에서 domain패키지로 이동
