+++ src/main/java/team/haedal/gifticionfunding/core/jwt/JwtProvider.java
@@ -6,7 +6,7 @@ import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import team.haedal.gifticionfunding.domain.Role;
+import team.haedal.gifticionfunding.domain.user.Role;
 
 import javax.crypto.spec.SecretKeySpec;
 import java.security.Key;
