--- src/main/java/team/haedal/gifticionfunding/domain/Vendor.java
@@ -1,5 +1,0 @@
-package team.haedal.gifticionfunding.domain;
-
-public enum Vendor {
-    KAKAO,
-}
