+++ src/main/java/team/haedal/gifticionfunding/domain/giftucon/GifticonSort.java
@@ -1,4 +1,4 @@
-package team.haedal.gifticionfunding.domain;
+package team.haedal.gifticionfunding.domain.giftucon;
 
 public enum GifticonSort {
     PRICE_ASC, PRICE_DESC, NAME_ASC, NAME_DESC
