+++ src/main/java/team/haedal/gifticionfunding/domain/user/Role.java
@@ -1,4 +1,4 @@
-package team.haedal.gifticionfunding.domain;
+package team.haedal.gifticionfunding.domain.user;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
