+++ src/main/java/team/haedal/gifticionfunding/domain/user/Vendor.java
@@ -1,0 +1,5 @@
+package team.haedal.gifticionfunding.domain.user;
+
+public enum Vendor {
+    KAKAO,
+}
