+++ src/main/java/team/haedal/gifticionfunding/repository/auth/OAuthClient.java
@@ -1,7 +1,7 @@
 package team.haedal.gifticionfunding.repository.auth;
 
-import team.haedal.gifticionfunding.domain.Vendor;
-import team.haedal.gifticionfunding.domain.VendorUserInfo;
+import team.haedal.gifticionfunding.domain.user.Vendor;
+import team.haedal.gifticionfunding.domain.user.VendorUserInfo;
 
 public interface OAuthClient {
     boolean canHandle(Vendor vendor);
