+++ src/main/java/team/haedal/gifticionfunding/repository/user/FriendshipActionJpaRepository.java
@@ -3,7 +3,7 @@ package team.haedal.gifticionfunding.repository.user;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import team.haedal.gifticionfunding.domain.FriendShipActionStatus;
+import team.haedal.gifticionfunding.domain.user.FriendShipActionStatus;
 import team.haedal.gifticionfunding.entity.user.FriendshipAction;
 
 public interface FriendshipActionJpaRepository extends JpaRepository<FriendshipAction, Long> {
