+++ src/test/java/team/haedal/gifticionfunding/service/gifticon/GifticonServiceTest.java
@@ -11,8 +11,8 @@ import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.test.context.support.WithMockUser;
-import team.haedal.gifticionfunding.domain.GifticonSearch;
-import team.haedal.gifticionfunding.domain.GifticonSort;
+import team.haedal.gifticionfunding.domain.giftucon.GifticonSearch;
+import team.haedal.gifticionfunding.domain.giftucon.GifticonSort;
 import team.haedal.gifticionfunding.domain.Status;
 import team.haedal.gifticionfunding.dto.common.PagingResponse;
 import team.haedal.gifticionfunding.dto.gifticon.response.GifticonModel;
